Refactor _formatScopeName to always operate on an array
It seems conceptually simpler to me to wrap an array around a single scope name at the beginning of the function. A `join` on an array with a single item just returns the one item, so the behavior is the same.

There is one behavior change, which is that `trim` was only being called when `_scopeName` was an array. I can't tell if this is an inconsistency or intentional, perhaps to allow padding spaces onto the end of the scope name to manually align multiple scopes? If that’s the case, the behavior will be obviated in a future commit.