Add an option for aligning scopes
Every time a new logger instance is created, it stores the scope(s) in a static (class) property. Then when the scopes are output, we can look through the array to find the longest scope(s) in order to figure out how much to pad the scope(s) by.

The alignment will happen on the “pointer”. So it will look something like this:

```
[first scope]  > output
[second scope] > output
```

#49